+++ CHANGELOG.md
@@ -2,6 +2,6 @@
 
 * [x] add option ```--data-extent``` to check data extent (and projection)
 * [x] add option ```--plugins``` and remove maven profile "cnig"
-* [ ] replace LegacyReportBuilder by JsonReportBuilder
 * [ ] simplify error context management
+* [ ] replace LegacyReportBuilder by JsonReportBuilder
 * [ ] log & check ogr2ogr's version and ban some versions
