+++ validator-core/src/main/java/fr/ign/validator/model/type/GeometryType.java
@@ -4,9 +4,9 @@ import com.vividsolutions.jts.geom.Geometry;
 import com.vividsolutions.jts.geom.GeometryCollection;
 import com.vividsolutions.jts.io.WKTReader;
 
-import fr.ign.validator.model.Attribute;
+import fr.ign.validator.data.Attribute;
 import fr.ign.validator.model.AttributeType;
-import fr.ign.validator.validation.GeometryIsValidValidator;
+import fr.ign.validator.validation.attribute.GeometryIsValidValidator;
 
 public class GeometryType extends AttributeType<Geometry> {
