+++ validator-core/src/main/java/fr/ign/validator/model/type/IntegerType.java
@@ -1,6 +1,6 @@
 package fr.ign.validator.model.type;
 
-import fr.ign.validator.model.Attribute;
+import fr.ign.validator.data.Attribute;
 import fr.ign.validator.model.AttributeType;
 
 public class IntegerType extends AttributeType<Integer> {
