+++ validator-core/src/main/java/fr/ign/validator/process/CharsetPreProcess.java
@@ -14,9 +14,9 @@ import org.jdom.JDOMException;
 
 import fr.ign.validator.Context;
 import fr.ign.validator.ValidatorListener;
+import fr.ign.validator.data.Document;
+import fr.ign.validator.data.DocumentFile;
 import fr.ign.validator.error.ErrorCode;
-import fr.ign.validator.model.Document;
-import fr.ign.validator.model.DocumentFile;
 import fr.ign.validator.model.FileModel;
 import fr.ign.validator.model.file.MetadataModel;
 import fr.ign.validator.reader.MetadataReader;
