+++ validator-core/src/main/java/fr/ign/validator/process/NormalizePostProcess.java
@@ -29,9 +29,9 @@ import com.vividsolutions.jts.geom.Geometry;
 
 import fr.ign.validator.Context;
 import fr.ign.validator.ValidatorListener;
+import fr.ign.validator.data.Document;
+import fr.ign.validator.data.DocumentFile;
 import fr.ign.validator.model.AttributeType;
-import fr.ign.validator.model.Document;
-import fr.ign.validator.model.DocumentFile;
 import fr.ign.validator.model.FeatureType;
 import fr.ign.validator.model.FileModel;
 import fr.ign.validator.model.file.MetadataModel;
