+++ validator-core/src/main/java/fr/ign/validator/validation/attribute/GeometryDataExtentValidator.java
@@ -1,11 +1,11 @@
-package fr.ign.validator.validation;
+package fr.ign.validator.validation.attribute;
 
 import com.vividsolutions.jts.geom.Geometry;
 
 import fr.ign.validator.Context;
+import fr.ign.validator.data.Attribute;
 import fr.ign.validator.error.ErrorCode;
-import fr.ign.validator.model.Attribute;
-import fr.ign.validator.model.Validator;
+import fr.ign.validator.validation.Validator;
 
 /**
  *
