+++ validator-core/src/main/java/fr/ign/validator/validation/document/DocumentDirectoryNameValidator.java
@@ -1,9 +1,9 @@
-package fr.ign.validator.validation;
+package fr.ign.validator.validation.document;
 
 import fr.ign.validator.Context;
+import fr.ign.validator.data.Document;
 import fr.ign.validator.error.ErrorCode;
-import fr.ign.validator.model.Document;
-import fr.ign.validator.model.Validator;
+import fr.ign.validator.validation.Validator;
 
 /**
  *
