+++ validator-plugin-cnig/src/main/java/fr/ign/validator/cnig/info/DocumentInfoExtractor.java
@@ -22,9 +22,9 @@ import fr.ign.validator.Context;
 import fr.ign.validator.cnig.utils.EnveloppeUtils;
 import fr.ign.validator.cnig.utils.IdurbaUtils;
 import fr.ign.validator.cnig.utils.TyperefExtractor;
+import fr.ign.validator.data.Document;
+import fr.ign.validator.data.DocumentFile;
 import fr.ign.validator.error.ErrorCode;
-import fr.ign.validator.model.Document;
-import fr.ign.validator.model.DocumentFile;
 import fr.ign.validator.model.file.MetadataModel;
 import fr.ign.validator.reader.MetadataReader;
 import fr.ign.validator.tools.CompanionFileUtils;
