+++ validator-plugin-cnig/src/main/java/fr/ign/validator/cnig/process/CreateShapefilesPostProcess.java
@@ -12,7 +12,7 @@ import org.apache.logging.log4j.MarkerManager;
 
 import fr.ign.validator.Context;
 import fr.ign.validator.ValidatorListener;
-import fr.ign.validator.model.Document;
+import fr.ign.validator.data.Document;
 import fr.ign.validator.model.FeatureType;
 import fr.ign.validator.tools.FileConverter;
 import fr.ign.validator.tools.VRT;
