+++ validator-plugin-cnig/src/main/java/fr/ign/validator/cnig/validation/AtLeastOneWritingMaterialValidator.java
@@ -7,11 +7,11 @@ import org.apache.logging.log4j.MarkerManager;
 
 import fr.ign.validator.Context;
 import fr.ign.validator.ValidatorListener;
+import fr.ign.validator.data.Document;
+import fr.ign.validator.data.DocumentFile;
 import fr.ign.validator.error.ErrorCode;
-import fr.ign.validator.model.Document;
-import fr.ign.validator.model.DocumentFile;
-import fr.ign.validator.model.Validator;
 import fr.ign.validator.model.file.PdfModel;
+import fr.ign.validator.validation.Validator;
 
 /**
  *
