+++ validator-plugin-cnig/src/main/java/fr/ign/validator/cnig/validation/InseeValidator.java
@@ -10,15 +10,15 @@ import org.apache.logging.log4j.MarkerManager;
 import fr.ign.validator.Context;
 import fr.ign.validator.ValidatorListener;
 import fr.ign.validator.cnig.utils.InseeUtils;
+import fr.ign.validator.data.Attribute;
+import fr.ign.validator.data.Document;
 import fr.ign.validator.error.ErrorCode;
-import fr.ign.validator.model.Attribute;
 import fr.ign.validator.model.AttributeType;
-import fr.ign.validator.model.Document;
 import fr.ign.validator.model.FeatureType;
 import fr.ign.validator.model.FileModel;
-import fr.ign.validator.model.Validator;
 import fr.ign.validator.model.file.TableModel;
 import fr.ign.validator.model.type.StringType;
+import fr.ign.validator.validation.Validator;
 
 /**
  * Extension du validateur pour valider les codes INSEE
